\name{coerce-methods}
\docType{methods}
\alias{coerce-methods}
\alias{coerce,groupedData,data.frame}
\alias{coerce,lme,reStruct-method}
\alias{coerce,matrix,pdmatrix-method}
\alias{coerce,pdCompSymm,pdmatrix-method}
\alias{coerce,pdDiag,pdmatrix-method}
\alias{coerce,pdIdent,corrmatrix-method}
\alias{coerce,pdIdent,pdmatrix-method}
\alias{coerce,pdLogChol,pdmatrix-method}
\alias{coerce,pdMat,corrmatrix-method}
\alias{coerce,pdMat,matrix-method}
\alias{coerce,pdNatural,corrmatrix-method}
\alias{coerce,pdNatural,pdmatrix-method}
\alias{coerce,pdfactor,pdmatrix-method}
\alias{coerce,pdmatrix,corrmatrix-method}
\alias{coerce,pdmatrix,pdfactor-method}
\alias{coerce<-,lme,reStruct-method}
\alias{coerce<-,pdDiag,pdmatrix-method}
\alias{coerce<-,pdIdent,pdmatrix-method}
\alias{coerce<-,pdLogChol,pdmatrix-method}
\alias{coerce<-,pdMat,matrix-method}
\alias{coerce<-,pdMat,pdfactor-method}
\alias{coerce<-,pdNatural,corrmatrix-method}
\alias{coerce<-,pdNatural,pdmatrix-method}
\alias{coerce<-,pdmatrix,pdfactor-method}
\title{Coercion methods and assignments}
\description{Methods for \code{coerce} or \code{"coerce<-"} are
  generated from \code{setAs} relationships between classes.  Typically
  these methods are invoked indirectly.}
\keyword{manip}
\keyword{methods}
