\name{densityplot.thpr}
\alias{densityplot.thpr}
\title{densityplot from a mixed-effects profile}
\usage{
  \method{densityplot}{thpr} (x, data, ...)
}
\arguments{
  \item{x}{a mixed-effects profile}

  \item{data}{not used - for compatibility with generic}

  \item{...}{optional arguments to
  \code{\link[lattice]{densityplot}()} from package
  \pkg{lattice}.}
}
\value{
  a density plot
}
\description{
  Densityplot method for a mixed-effects model profile
}
\examples{
## see   example("profile.merMod")
}

