\name{family-class}
\docType{class}
\alias{family-class}
\title{Class "family"}
\description{
  Family objects describe the error distributions and link functions
  used in generalized linear models and its extensions. It is used by
  functions such as \code{\link[base:glm]{glm}} and 
  \code{\link{GLMM}}.
}
\section{Objects from the Class}{This is an S3 style class. No objects
  may be created from it using \code{new}.}

\section{Extends}{
Class \code{"oldClass"}, directly.
}
\section{Methods}{
  No methods defined with class "family" in the signature.
}
%\references{}
%\author{}
%\note{}
%\seealso{}
%\examples{}
\keyword{classes}
