\name{getDoublevertDefault}
\alias{getDoublevertDefault}
\title{Retrieve default double-vertical-bar-expansion method}
\description{
  Retrieve information about whether random effect terms specified with
  double bars (\code{||}) should be split semantically into separate,
  intercept-free terms, or encoded with a \code{diag()} special
}
\usage{
getDoublevertDefault()
}
\value{
	a (1-element) character vector suitable for \code{reformulas::findbars_x}
	specifying whether to expand random-effects terms containing double bars (\code{||})
	as a set of independent random effects terms ("split": i.e., \code{(1+a+b||f)} becomes
	\code{(1|f) + (0+a|f) + (0+b|f)} or whether to specify that the covariance matrix is
	diagonal ("diag_special": i.e., \code{diag(1+a+b|f)}).

	"split" is backward compatible with the behaviour of \code{lme4} < 2.0; "diag_special"
	works properly/as expected with factors as varying terms. That is, if \code{g} is a factor
	then \code{(1 + g||f)} will expand (with method "split") to \code{(1|f) + (0 + g|f)}; the
	second term will \emph{not} produce a covariance matrix where the contrasts generated by
	the factor are independent.


}


