\name{getGrouping}
\alias{getGrouping}
\title{Extract the grouping factor from a random effect specification}
\usage{
  getGrouping(bar, fr)
}
\arguments{
  \item{bar}{the random effect specification as returned by
  \code{findbars}}

  \item{fr}{a \code{model.frame}}
}
\value{
  a factor
}
\description{
  Extract the grouping factor from a random effect
  specification
}

