\name{groupedData-class}
\docType{class}
\alias{groupedData-class}
\alias{formula,groupedData-method}
\title{Class "groupedData"}
\description{A \code{groupedData} object is at least a data.frame and a
  display  formula that provides a grouping structure.  At present we
  are using an old-style class for this.}
\section{Objects from the Class}{A virtual Class: No objects may be
  created from it.}
\section{Extends}{
Class \code{"data.frame"}, directly.
Class \code{"oldClass"}, by class "data.frame".
}
\section{Methods}{
  No methods defined with class "groupedData" in the signature.
}
%\references{}
%\author{}
%\note{}
%\seealso{}
%\examples{}
\keyword{classes}
