\name{mkLambdatUS}
\alias{mkLambdatUS}
\title{Initalize upper triangular Cholesky factor of an
unstructured covariance matrix}
\usage{
  mkLambdatUS(nc, nl, theta)
}
\arguments{
  \item{nc}{number of effects per level}

  \item{nl}{number of levels}

  \item{theta}{vector of variance-covariance parameters}
}
\value{
  a sparse upper triangular matrix
}
\description{
  Initalize upper triangular Cholesky factor of an
  unstructured covariance matrix
}

