\name{nffGroupedData-class}
\docType{class}
\alias{nffGroupedData-class}
\title{Class "nffGroupedData"}
\description{A \code{groupedData} object is at least a data.frame and a
  display  formula that provides a grouping structure.  In the
  \code{nfGroupedData} class there is only one level of grouping factors
  and the response in the display formula is numeric.  The
  \code{nffGroupedData} class is a further refinement that specifies
  that the primary covariate is a factor. At present we
  are using old-style classes for \code{groupedData} and the classes
  that extend it.}
\section{Objects from the Class}{A virtual Class: No objects may be
  created from it.}
\section{Extends}{
Class \code{"nfGroupedData"}, directly.
Class \code{"groupedData"}, by class "nfGroupedData".
Class \code{"data.frame"}, by class "nfGroupedData".
Class \code{"oldClass"}, by class "nfGroupedData".
}
\section{Methods}{
  No methods defined with class "nfnGroupedData" in the signature.
}
%\references{}
%\author{}
%\note{}
%\seealso{}
%\examples{}
\keyword{classes}
