\name{nlmeInternal-class}
\docType{class}
\alias{lmeLevelList-class}
\alias{summary.glmm-class}
\alias{summary.lme-class}
\alias{summary.pdMat-class}
\title{Some internal classes}
\description{These classes are used for internal purposes.}
%\references{}
%\author{}
%\note{}
%\examples{}
\keyword{classes}
