\name{nmGroupedData-class}
\docType{class}
\alias{nmGroupedData-class}
\title{Class "nmGroupedData"}
\description{A \code{groupedData} object is at least a data.frame and a
  display  formula that provides a grouping structure.  In the
  \code{nmGroupedData} class there are multiple level of grouping factors
  and the response in the display formula is numeric.  At present we
  use old-style classes for \code{groupedData} and the classes
  that inherit from it.}
\section{Objects from the Class}{A virtual Class: No objects may be
  created from it.}
\section{Extends}{
Class \code{"groupedData"}, directly.
Class \code{"data.frame"}, by class "groupedData".
Class \code{"oldClass"}, by class "groupedData".
}
\section{Methods}{
  No methods defined with class "groupedData" in the signature.
}
%\references{}
%\author{}
%\note{}
%\seealso{}
%\examples{}
\keyword{classes}
