\name{r5_to_list}
\alias{r5_to_list}
\title{Convert reference class objects to lists}
\usage{
  r5_to_list(obj)
}
\arguments{
  \item{obj}{An r5 object.}
}
\value{
  A list.
}
\description{
  Store the current return values of the fields of a
  reference class object in the elements of a list object.
}

