\docType{class}
\name{rePos-class}
\alias{rePos-class}
\title{Class \code{"rePos"}}
\description{
  A reference class for determining the positions in the
  random-effects vector that correspond to particular
  random-effects terms in the model formula
}
\section{Extends}{
  All reference classes extend and inherit methods from
  \code{"\linkS4class{envRefClass}"}.
}
\examples{
showClass("rePos")
}
\keyword{classes}

