\name{VirtualCovariance-class}
\alias{VirtualCovariance-class}
\alias{HomogeneousVariance-class}
\alias{HeterogeneousVariance-class}
\alias{UnstructuredCovariance-class}
\alias{DiagonalCovariance-class}
\alias{CSCovariance-class}
\alias{AR1Covariance-class}
\alias{HomogeneousDiagonalCovariance-class}
\alias{HeterogeneousDiagonalCovariance-class}
\alias{HomogeneousCSCovariance-class}
\alias{HeterogeneousCSCovariance-class}
\alias{HomogeneousAR1Covariance-class}
\alias{HeterogeneousAR1Covariance-class}
\alias{show,VirtualCovariance-method}
\docType{class}
\title{Virtual Class \sQuote{VirtualCovariance} of Covariance Matrices}
\description{
  Classes representing covariance matrices can be documented here.
}
\keyword{classes}
\keyword{internal}
