\name{glmFamily}
\alias{glmFamily}
\title{Generator object for the \code{\linkS4class{glmFamily}} class}
\section{Methods}{
  \describe{
    \item{\code{new(family=family)}:}{Create a new
      \code{\linkS4class{glmFamily}} object}
  }
}
\arguments{
  \item{family}{a \code{\link{family}} object}
}
\note{Arguments to the \code{new} method must be named arguments.}
\description{The generator object for the \code{\linkS4class{glmFamily}}
  reference class. Such an object is primarily used through its
  \code{new} method.}
\seealso{
  \code{\linkS4class{glmFamily}}
}
\keyword{classes}
