\name{glmFamily}
\alias{glmFamily}
\title{Generator object for the \code{\linkS4class{glmFamily}} class}
\usage{
  glmFamily(...)
}
\arguments{
  \item{...}{Named argument (see Note below)}
}
\description{
  The generator object for the
  \code{\linkS4class{glmFamily}} reference class. Such an
  object is primarily used through its \code{new} method.
}
\note{
  Arguments to the \code{new} method must be named
  arguments.
}
\section{Methods}{
  \describe{ \item{\code{new(family=family)}}{Create a new
  \code{\linkS4class{glmFamily}} object} }
}
\seealso{
  \code{\linkS4class{glmFamily}}
}
\keyword{classes}

