\docType{class}
\name{golden-class}
\alias{golden-class}
\title{Class \code{"golden"}}
\description{
  A reference class for a golden search scalar optimizer
  using reverse communication.
}
\section{Extends}{
  All reference classes extend and inherit methods from
  \code{"\linkS4class{envRefClass}"}.
}
\examples{
showClass("golden")
}
\keyword{classes}

