\docType{class}
\name{lmList-class}
\alias{lmList-class}
\alias{show,lmList-method}
\title{Class "lmList" of 'lm' Objects on Common Model}
\description{
  Class \code{"lmList"} is an S4 class with basically a
  list of objects of class \code{\link{lm}} with a common
  model.
}
\section{Objects from the Class}{
  Objects can be created by calls of the form
  \code{new("lmList", ...)} or, more commonly, by a call to
  \code{\link{lmList}}.
}
\keyword{classes}

