\name{lFormula}
\alias{glFormula}
\alias{lFormula}
\alias{mkGlmerDevfun}
\alias{mkLmerDevfun}
\alias{optimizeGlmer}
\alias{optimizeLmer}
\alias{reoptimizeGlmer}
\alias{updateGlmerDevfun}
\title{Modular functions for mixed model fits}
\usage{
  lFormula(formula, data = NULL, REML = TRUE,
    sparseX = FALSE, control = list(), start = NULL,
    verbose = 0L, subset, weights, na.action, offset,
    contrasts = NULL, devFunOnly = FALSE,
    optimizer = "Nelder_Mead", ...)

  mkLmerDevfun(fr, X, reTrms, REML = TRUE, start = NULL,
    ...)

  optimizeLmer(devfun, reTrms, optimizer = "Nelder_Mead",
    verbose = 0L, control = list())

  glFormula(formula, data = NULL, family = gaussian,
    sparseX = FALSE, control = list(), start = NULL,
    verbose = 0L, nAGQ = 1L, compDev = TRUE, subset,
    weights, na.action, offset, contrasts = NULL, mustart,
    etastart, devFunOnly = FALSE, tolPwrss = 1e-07,
    optimizer = c("bobyqa", "Nelder_Mead"), ...)

  mkGlmerDevfun(fr, X, reTrms, family, nAGQ = 1L,
    verbose = 0L, tolPwrss = 1e-07, compDev = TRUE)

  optimizeGlmer(devfun, optimizer = "bobyqa", nAGQ = 1L,
    verbose = 0L, control = list())

  updateGlmerDevfun(devfun, reTrms, nAGQ = 1L)

  reoptimizeGlmer(devfun, optimizer = "Nelder_Mead",
    verbose = 0L, control = list())
}
\arguments{
  \item{\dots}{other potential arguments.}

  \item{fr}{model frame}

  \item{X}{fixed-effects design matrix}

  \item{reTrms}{information on random effects structure
  (see \code{\link{mkReTrms}})}

  \item{REML}{(logical) fit restricted maximum likelihood
  model?}

  \item{start}{starting values}

  \item{devfun}{a deviance function, as generated by
  \code{\link{mkLmerDevfun}}}

  \item{nAGQ}{number of Gauss-Hermite quadrature points}

  \item{formula}{a two-sided linear formula object
  describing both the fixed-effects and fixed-effects part
  of the model, with the response on the left of a \code{~}
  operator and the terms, separated by \code{+} operators,
  on the right.  Random-effects terms are distinguished by
  vertical bars (\code{"|"}) separating expressions for
  design matrices from grouping factors.}

  \item{data}{an optional data frame containing the
  variables named in \code{formula}.  By default the
  variables are taken from the environment from which
  \code{lmer} is called. While \code{data} is optional, the
  package authors \emph{strongly} recommend its use,
  especially when later applying methods such as
  \code{update} and \code{drop1} to the fitted model
  (\emph{such methods are not guaranteed to work properly
  if \code{data} is omitted}). If \code{data} is omitted,
  variables will be taken from the environment of
  \code{formula} (if specified as a formula) or from the
  parent frame (if specified as a character vector).}

  \item{sparseX}{logical - should a sparse model matrix be
  used for the fixed-effects terms?  Defaults to
  \code{FALSE}. Currently inactive.}

  \item{control}{a named list of control parameters for the
  estimation algorithm, specifying only the ones to be
  changed from their default values.  Hence defaults to an
  empty list.\cr Possible control options and their default
  values are: \describe{ \item{\code{msVerbose}:}{a logical
  value passed as the \code{trace} argument to
  \code{nlminb} (see documentation on that function).
  Default is \code{getOption("verbose")}.}
  \item{\code{maxIter}:}{a positive integer passed as the
  \code{maxIter} argument to \code{nlminb} (see
  documentation on that function).  Default is \code{300}.}
  \item{\code{maxFN}:}{a positive integer specifying the
  maximum number of evaluations of the deviance function
  allowed during the optimization. Default is \code{900}.}
  \item{\code{tol}:}{a positive number specifying the
  convergence tolerance, currently only for the PWRSS
  iterations in \code{\link{glmer}}.  Default is
  \code{0.000001}.} }}

  \item{verbose}{integer scalar.  If \code{> 0} verbose
  output is generated during the optimization of the
  parameter estimates.  If \code{> 1} verbose output is
  generated during the individual PIRLS steps.}

  \item{subset}{an optional expression indicating the
  subset of the rows of \code{data} that should be used in
  the fit. This can be a logical vector, or a numeric
  vector indicating which observation numbers are to be
  included, or a character vector of the row names to be
  included.  All observations are included by default.}

  \item{weights}{an optional vector of \sQuote{prior
  weights} to be used in the fitting process.  Should be
  \code{NULL} or a numeric vector.}

  \item{na.action}{a function that indicates what should
  happen when the data contain \code{NA}s.  The default
  action (\code{na.fail}) prints an error message and
  terminates if there are any incomplete observations.}

  \item{offset}{this can be used to specify an \emph{a
  priori} known component to be included in the linear
  predictor during fitting. This should be \code{NULL} or a
  numeric vector of length equal to the number of cases.
  One or more \code{\link{offset}} terms can be included in
  the formula instead or as well, and if more than one is
  specified their sum is used.  See
  \code{\link{model.offset}}.}

  \item{contrasts}{an optional list. See the
  \code{contrasts.arg} of \code{model.matrix.default}.}

  \item{devFunOnly}{logical - return only the deviance
  evaluation function.}

  \item{optimizer}{character - name of optimizing function.
  The built-in optimizers are \code{\link{Nelder_Mead}} and
  \code{\link[minqa]{bobyqa}} (from the \pkg{minqa}
  package. Any minimizing function that allows box
  constraints can be used provided that it (1) takes input
  parameters \code{fn} (function to be optimized),
  \code{par} (starting parameter values), \code{lower}
  (lower bounds) and \code{control} (control parameters,
  passed through from the \code{control} argument) and (2)
  returns a list with (at least) elements \code{par}
  (best-fit parameters), \code{fval} (best-fit function
  value), \code{conv} (convergence code) and (optionally)
  \code{message} (informational message, or explanation of
  convergence failure).  Special provisions are made for
  \code{\link{bobyqa}}, \code{\link{Nelder_Mead}}, and
  optimizers wrapped in the \pkg{optimx} package; to use
  \pkg{optimx} optimizers (including \code{L-BFGS-B} from
  base \code{\link{optim}} and \code{\link{nlminb}}), pass
  the \code{method} argument to \code{optim} in the
  \code{control} argument.}

  \item{family}{a GLM family, see \code{\link[stats]{glm}}
  and \code{\link[stats]{family}}.}

  \item{compDev}{logical scalar - should compiled code be
  used for the deviance evaluation during the optimization
  of the parameter estimates?  Defaults to \code{TRUE}.}

  \item{mustart}{optional starting values on the scale of
  the conditional mean, as in \code{\link[stats]{glm}}; see
  there for details.}

  \item{etastart}{optional starting values on the scale of
  the unbounded predictor as in \code{\link[stats]{glm}};
  see there for details.}

  \item{tolPwrss}{numeric scalar - the tolerance for
  declaring convergence in the penalized iteratively
  weighted residual sum-of-squares step.  Defaults to
  1e-10.}
}
\value{
  a list containing components: \item{fr}{model frame}
  \item{X}{fixed-effect design matrix} \item{reTrms}{list
  containing information on random effects structure:
  result of \code{\link{mkMerMod}}} \item{REML}{logical
  flag: use restricted maximum likelihood? (Copy of
  argument.)} \item{start}{starting value information.
  (Copy of argument.)}
  a function to calculate deviance (or restricted deviance)
  as a function of the theta (random-effect) parameters
  results of an optimization
}
\description{
  Modular functions for mixed model fits

  process
}
\details{
  Takes all the end-user-level components required for
  fitting a linear mixed model and produces a list of terms
  required for the next internal stage in fitting
  (\code{\link{mkLmerDevfun}})
}

