\name{nlmeInternal-class}
\docType{class}
\alias{lmeLevelList-class}
\alias{summary.glmm-class}
\alias{show,summary.glmm-method}
\alias{summary.lme-class}
\alias{show,summary.lme-method}
\alias{summary.pdMat-class}
\title{Some internal classes}
\description{These classes are used for internal use only.  When
  namespaces are available for packages that use S4 classes and methods
  these classes will be hidden inside the namespace}
%\references{}
%\author{}
%\note{}
%\examples{}
\keyword{classes}
